+++ src/basic/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './use-local-storage';
 export * from './use-notifications';
+export * from './use-cart';
